refactor/frontend: move mockEventClass.js to src/test
We already have a test utility folder,
and there the import looks nicer.
