+++ frontend/src/components/form/api/__tests__/ApiTextarea.spec.js
@@ -9,7 +9,7 @@ import { ApiMock } from '@/components/form/api/__tests__/ApiMock'
 import { i18n } from '@/plugins'
 import { mount as mountComponent } from '@vue/test-utils'
 import { waitForDebounce } from '@/test/util'
-import { mockEventClass } from '@/../tests/mockEventClass'
+import { mockEventClass } from '@/test/mockEventClass'
 
 Vue.use(Vuetify)
 Vue.use(formBaseComponents)
