+++ frontend/src/components/form/base/__tests__/ETextArea.spec.js
@@ -5,7 +5,7 @@ import formBaseComponents from '@/plugins/formBaseComponents'
 
 import { mount as mountComponent } from '@vue/test-utils'
 import ETextarea from '../ETextarea.vue'
-import { mockEventClass } from '@/../tests/mockEventClass'
+import { mockEventClass } from '@/test/mockEventClass'
 
 Vue.use(Vuetify)
 Vue.use(formBaseComponents)
