+++ frontend/src/test/mockEventClass.js
